+++ docs/changelog.md
@@ -1,4 +1,9 @@
 # TBD
+### Features
+* Upgrade to `module-api-lib` 0.12.0, which is compatible with the latest version of Kurtosis
+
+### Removals
+* Removed the unused `nearup` service
 
 # 0.5.10
 ### Fixes